Unify flush and compaction code paths
Remove `DB.writeLevel0Table` and use a shared code path for flushes and
compactions. There was significant overlap in these previously separate
code paths that this makes sense to do purely from a cleanliness point
of view. Additionally, it found a bug in compactions where a compaction
that was empty save for a range tombstone was silently dropping that
tombstone (flushes were previously handling this case correctly, now the
shared code path does too). This paves the way for experimenting with
flushes that also compact with L0 or other levels and for flushes that
create more than 1 output table.